Move files to classes / book file
