+++ classes/book/label.rb
